# backend/api.yaml
openapi: "3.0.0"
info:
  title: Chaos API
  version: 1.0.0
servers:
  - url: https://chaos.csesoc.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local server
paths:
  /auth/callback/google:
    get:
      operationId: googleCallback
      description: Google OAuth callback
      tags:
        - Auth
      parameters:
        - name: code
          in: query
          description: Google OAuth code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: string
                description: JWT token
                example: "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9..."
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotLoggedIn"
  /auth/logout:
    post:
      operationId: logout
      description: Invalidates current token
      tags:
        - Auth
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Successfully logged out
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotLoggedIn"
  /user:
    get:
      operationId: getLoggedInUser
      description: Returns info about currently logged in user
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotLoggedIn"
  /user/name:
    patch:
      operationId: updateUserName
      description: Updates currently logged in user's name
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: "Clancy Tiger"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Updated username
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /user/pronouns:
    patch:
      operationId: updateUserPronouns
      description: Updates currently logged in user's pronouns
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                pronouns:
                  type: string
                  example: They/Them
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Updated pronouns
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /user/gender:
    patch:
      operationId: updateUserGender
      description: Updates currently logged in user's gender
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                gender:
                  type: string
                  example: Female
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Updated gender
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /user/zid:
    patch:
      operationId: updateUserZid
      description: Updates currently logged in user's zID
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                zid:
                  type: string
                  example: z5123456
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Updated zid
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /user/degree:
    patch:
      operationId: updateUserDegree
      description: Updates currently logged in user's degree
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                degree_name:
                  type: string
                  example: Electrical Engineering
                degree_starting_year:
                  type: integer
                  example: 2024
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Updated degree
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /user/applications:
    get:
      operationId: getUserApplications
      description: Returns info about applications made by currently logged in user
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationDetails"
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation:
    post:
      operationId: createOrganisation
      description: Create a new organisation
      tags:
        - Organisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOrganisation"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Successfully created organisation
        "401":
          $ref: "#/components/responses/NotSuperUser"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/slug_check:
    post:
      operationId: checkOrganisationSlugAvailability
      description: Check if slug is available
      tags:
        - Organisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlugCheck"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Organisation slug is available
        "400":
          description: Bad request - slug is in use or not ASCII
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
        "401":
          $ref: "#/components/responses/NotSuperUser"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}:
    get:
      operationId: getOrganisationById
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Returns info about specified organisation
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationDetails"
        "401":
          $ref: "#/components/responses/NotLoggedIn"
    delete:
      operationId: deleteOrganisationById
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Deletes specified organisation
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Successfully deleted organisation
        "401":
          $ref: "#/components/responses/NotSuperUser"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/slug/{slug}:
    get:
      operationId: getOrganisationBySlug
      parameters:
        - name: slug
          in: path
          description: Organisation slug
          required: true
          schema:
            type: string
      description: Returns info about specified organisation
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationDetails"
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}/campaign:
    post:
      operationId: createCampaign
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Create a new campaign inside specified organisation
      tags:
        - Organisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCampaign"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Successfully created campaign
        "401":
          $ref: "#/components/responses/NotOrganisationAdmin"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}/campaign/slug_check:
    post:
      operationId: checkCampaignSlugAvailability
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Checks availability of campaign slug in specified organisation
      tags:
        - Organisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlugCheck"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Campaign slug is available
        "401":
          $ref: "#/components/responses/NotOrganisationAdmin"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}/campaigns:
    get:
      operationId: getAllOrganisationCampaigns
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Returns all campaigns for specified organisation
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganisationCampaign"
        "401":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}/email_template:
    post:
      operationId: createEmailTemplate
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Create a new email template within the organisation
      tags:
        - Organisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEmailTemplate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Successfully created email template
        "401":
          $ref: "#/components/responses/NotOrganisationAdmin"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}/email_templates:
    get:
      operationId: getAllOrganisationEmailTemplates
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Get all email templates for specified organisation
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
        "401":
          $ref: "#/components/responses/NotOrganisationAdmin"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
  /organisation/{id}/logo:
    patch:
      operationId: updateOrganisationLogoById
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      description: Update logo for specified organisation. Returns a PUT url to upload new image to
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  upload_url:
                    type: string
                    description: Presigned S3 url to upload file
                    example: https://presignedurldemo.s3.eu-west-2.amazonaws.com/6996987893965262849/2d19617b-46fd-4927-9f53-77d69232ba5d
        "401":
          $ref: "#/components/responses/NotOrganisationAdmin"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
...
  /organisation/{id}/member:
    delete:
      operationId: deleteOrganisationMember
      parameters:
        - name: id
          in: path
          description: Organisation ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                user_id:
                  type: integer
                  format: int64
      description: Removes a member from the organisation
      tags:
        - Organisation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Successfully removed member
        "401":
          $ref: "#/components/responses/NotOrganisationAdmin"
        "307":
          $ref: "#/components/responses/NotLoggedIn"
...
components:
  schemas:
    NotLoggedIn:
      type: object
      properties:
        message:
          type: string
          example: Not logged in
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1541815603606036480
        email:
          type: string
          example: clancy.lion@unsw.edu.au
        pronouns:
          type: string
          nullable: true
          example: he/him
        gender:
          type: string
          nullable: true
          example: male
        zid:
          type: string
          nullable: true
          example: z1234567
        degree:
          type: string
          nullable: true
          example: Computer Science
        created_at:
          type: string
          example: 2024-03-15T18:25:43.511Z
        updated_at:
          type: string
          example: 2024-03-15T18:25:43.511Z
...
    NewOrganisation:
      type: object
      properties:
        slug:
          type: string
          example: unsw-devsoc
        name:
          type: string
          example: UNSW Software Development Society
        admin:
          type: integer
          format: int64
          example: 1541815603606036480
    SlugCheck:
      type: object
      properties:
        slug:
          type: string
          example: unsw-devsoc
...
  responses:
    NotLoggedIn:
      description: Not logged in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotLoggedIn"
    NotSuperUser:
      description: User is not a super user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Unauthorized"
...